requirements.txt
python-telegram-bot==21.*
uvicorn==0.30.*
starlette==0.37.*